consolidate report-related actions
